refactor: Navbar 컴포넌트를 Header 컴포넌트로 수정
